Split database integration tests
